day06_llmops_monitoring/requirements.txt
pandas>=2.2.0
numpy>=1.26.0
evidently>=0.4.37
mlflow>=2.14.0
matplotlib>=3.8.0
scikit-learn>=1.4.0
pyyaml>=6.0
seaborn>=0.12.0
plotly>=5.20.0
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
